fix: move verifications to settings
move mobile verification to settings
disabled email verification
