+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "swash-extension",
-  "version": "2.13.0",
+  "version": "2.14.0",
   "type": "module",
   "repository": {
     "url": "https://github.com/swashapp/extension.git"
